+++ CHANGELOG.md
@@ -15,3 +15,5 @@
 - Shaders
 - Cameras
 - Lenses
+- UI components:
+  - `GlimpseSurfaceView` (Android)